Remove the ScrollView and force portrait mode
This commit would remove the ScrollView and force portrait mode, and also
align the clear button to the bottom of the screen, while aligning the five
card hand and the card selection area above the clear button.  These
changes would afford more empty space at the top of the screen for more
verbose text output.