Add serialized HashMap object to app and use this in calculating cards to hold.
This commit would add in a serialized HashMap file to the app, which
contains key-value pairs that would allow the user to determine the best choices of cards to hold on a five-card draw in video poker.  The commit
adds additional TextViews that display over the card images determined as
good choices for holds.  Large heapsize would also be enabled in the
manifest file.