Add logic to automatically label all cards in hand as hold if the hand contains four-of-a-kind or a full-house.
This commit would compensate for errors contained in the serialized object
that did not contain correct values for certain hands containing
four-of-a-kind.  The added logic in this commit would circumvent those
circumstances by imposing the decision to hold all cards in such hands,
regardless of the values contained in the corresponding HashMap entries.